Update: Move remaining interfaces to client
