+++ packages/client/src/utils/parameters.ts
@@ -1,17 +1,17 @@
 /**
- * Copyright 2020, Yahoo Holdings Inc.
+ * Copyright 2022, Yahoo Holdings Inc.
  * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
  */
-import FunctionParameterMetadataModel, { DataType } from '@yavin/client/models/metadata/function-parameter';
-import { isPresent } from '@ember/utils';
-import type { ParameterValue } from '@yavin/client/request';
+import { DataType } from '../models/metadata/function-parameter.js';
+import type FunctionParameterMetadataModel from '../models/metadata/function-parameter.js';
+import type { ParameterValue } from '../request.js';
 
 export function parseParameterValue(
   parameter: FunctionParameterMetadataModel,
   rawParamValue: ParameterValue
 ): ParameterValue {
   const { defaultValue, valueType } = parameter;
-  let paramValue: ParameterValue = isPresent(rawParamValue) ? rawParamValue : defaultValue ?? '';
+  let paramValue: ParameterValue = rawParamValue ? rawParamValue : defaultValue ?? '';
   if ([DataType.INTEGER, DataType.DECIMAL].includes(valueType)) {
     paramValue = Number(rawParamValue);
   } else if (DataType.BOOLEAN === valueType) {
