+++ packages/data/addon/adapters/dimensions/bard.ts
@@ -15,9 +15,8 @@ import { taskFor } from 'ember-concurrency-ts';
 import type { TaskGenerator } from 'ember-concurrency';
 import type { Filter } from '@yavin/client/request';
 import type NaviMetadataService from 'navi-data/services/navi-metadata';
-import type NaviDimensionAdapter from './interface';
-import type { DimensionFilter } from './interface';
-import type { Options } from './interface';
+import type NaviDimensionAdapter from '@yavin/client/adapters/dimensions/interface';
+import type { DimensionFilter, Options } from '@yavin/client/adapters/dimensions/interface';
 import type DimensionMetadataModel from '@yavin/client/models/metadata/dimension';
 import type { DimensionColumn } from '@yavin/client/models/metadata/dimension';
 import { searchDimensionRecords } from '@yavin/client/utils/search';
