+++ packages/data/addon/adapters/dimensions/elide.ts
@@ -14,9 +14,8 @@ import type { TaskGenerator } from 'ember-concurrency';
 import type { AsyncQueryResponse } from '@yavin/client/adapters/facts/interface';
 import type { FilterOperator } from '@yavin/client/request';
 import type { Column, RequestV2 } from '@yavin/client/request';
-import type NaviDimensionAdapter from './interface';
-import type { DimensionFilter } from './interface';
-import type { Options } from './interface';
+import type NaviDimensionAdapter from '@yavin/client/adapters/dimensions/interface';
+import type { DimensionFilter, Options } from '@yavin/client/adapters/dimensions/interface';
 import type ElideFactsAdapter from '../facts/elide';
 import type { DimensionColumn } from '@yavin/client/models/metadata/dimension';
 import type ElideDimensionMetadataModel from '@yavin/client/models/metadata/elide/dimension';
