--- packages/data/app/utils/dimension.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-data/utils/dimension';
